Changed the type from VersionData<T> to IPersistanceData<T> in IProcessManagerFinder
To allow us to add meta data to IProcessManagerData we used VersionData that wrapped IProcessManagerData and also had properties for version and id.  The problem is VersionData is specific to the MongoDB implementation of the Process Manager finder.  A more generic solution is to change the VersionData type to IPersistanceData.  Each Process Manager finder can then use its own type that implements the IPersistanceData interface and can add any properties it needs.  This solution means that all Process Manager finders dont need to use the version property that was on VersionData.