Added logic to the bus that starts, continues and removes process managers
If the message being received starts process managers a new process manager is created, executed and persisted.  If the message received is a handler inside a process manager the process manager is found using the process managers FindProcessManagerData method, the handler is executed and the data is either updated or deleted if it is complete.