Setting QueueName still does not work
The previous fix caused the default queue name to always override  the queue name set in a configuration file.
To fix it, I created a new 'temp' queue name variable  that is assigned when  the queue name is set via fluent API.
The queue name can be assigned from a configuration file only when the 'temp' queue name
(rather than 'default' queue name) is null or empty.